refactor: centralize wallet selector
